+++ examples/with-next-chakra-ui/next.config.js
@@ -1,15 +1,7 @@
-const withTM = require("next-transpile-modules")([
-  "d3-format",
-  "@wizard-ui/core",
-  "@wizard-ui/react",
-]);
-
 /** @type {import('next').NextConfig} */
 const config = {
-  reactStrictMode: false,
-  experimental: {
-    esmExternals: "loose",
-  },
+  reactStrictMode: true,
+  swcMinify: true,
 };
 
-module.exports = withTM(config);
+module.exports = config;