+++ packages/core/src/wallet/errors.ts
@@ -70,3 +70,7 @@ export class WalletWindowBlockedError extends WalletError {
 export class WalletWindowClosedError extends WalletError {
   name = "WalletWindowClosedError";
 }
+
+export class WalletNotSelectedError extends WalletError {
+  name = "WalletNotSelectedError";
+}