+++ packages/react/src/components/index.ts
@@ -1,6 +1,5 @@
 export * from "./WalletIcon";
 export * from "./WalletModal";
 export * from "./WalletModalProvider";
-export * from "./WalletProvider";
 export * from "./WalletButton";
 export * from "./WizardProvider";