+++ packages/react/src/errors.ts
@@ -1,5 +1,0 @@
-import { WalletError } from "@wizard-ui/core";
-
-export class WalletNotSelectedError extends WalletError {
-  name = "WalletNotSelectedError";
-}