+++ packages/react/src/hooks/index.ts
@@ -1,4 +1,3 @@
 export * from "./useExecuteContract";
 export * from "./useQueryContract";
 export * from "./useWalletModal";
-export * from "./useWallet";