+++ packages/react/src/index.ts
@@ -1,3 +1,3 @@
-export * from "./errors";
 export * from "./components";
+export * from "./contexts";
 export * from "./hooks";